fix
Переименование Data.cpp в SearchData.cpp и перенос реализаций классов SearchInterval, SearchIntervalFactory и SearchDataIterotor из SearchData.cpp в отдельный файлы.
